Fix crash from null star
An issue could occur when findStar() failed to find a star, causing the game to crash. I'm not sure why this happens, but adding a check for a null pointer seems to fix the issue.

Also while we're here, cleanup some unnecessary repeated calls to findSpaceWorm() and findStar() in this function.