Game over and return to menu
There is now a button when the game is paused to exit to the main menu. The
game will also now display a special game over screen when the player loses the
game showing them their final score and prompting them to play again or exit.

Fix a small issue where the invisibility powerup would tick down while the game
is paused.

Also I accidentally reformatted the entire SnakeGame.java file and didn't
realize until later, so there's a bunch of small changes in here due to that.