High scores functionality
Create a view on the high scores menu to display scores information, including the points received and the date. Though since we don't actually save this data including the date is more of a proof of concept than anything. The menu currently has no limit on the amount of scores, but could easily be changed to limit it to 5 or 10, for example.

Also includes some smaller changes like fixing an intermittent crash from drawing asteroids and calling the worm's updateInvisible method twice per frame.