Split scores by difficulty
Scores are now stored in separate lists according to the difficulties they were
obtained on. The scores menu now has three buttons to select which difficulty
to view scores for.

Also add some simple fade transitions between menu elements. Animating between
the main menu and the game SurfaceView unfortunately is not so simple because
it is running in a separate thread, so things end up looking a little weird.